New data layout vertices (#148)
* first version

* fix some performance issues

* remove EXADEM_MAX_VERTICES

* fix issues with rcut max

* rename some data structures used to store vertices

* reactive paraview output files

* test

* add example (polyhedron/stl_cylinder) + fix multishape that did not correctly work for different number of vertices

---------

+++ src/field/set_fields.cu
@@ -173,7 +173,7 @@ namespace exaDEM
           lout << std::endl;
           std::exit(EXIT_FAILURE);  
         }
-        int64_t type_id = type_map.at(type_name);
+        size_t type_id = type_map.at(type_name);
         // default values;
         double vx = 0;
         double vy = 0;
